==> scripts/04-data_simulation.qmd <==
---
title: "Simulating Data from the Toronto Paramedic Service"
format: html
author: "Sebastian Rodriguez"
---
```{r}
#### Preamble ####
# Purpose: Read in data from the Toronto Paramedic Service
# Prerequisites: Need to know where to get Toronto Paramedic Services data
# Author: Sebastian Rodriguez
# Email: me@srod.ca
# Date: 3 February 2023
# GitHub: https://github.com/seb646/toronto-paramedic-responses

#### Workspace set-up ####
library(tidyverse) # A collection of data-related packages
library(janitor) # Helps clean datasets
```

```{r}
#### Simulate data based on Land Ambulance Response Time Standard ####
# Link: https://open.toronto.ca/dataset/land-ambulance-response-time-standard 
simulated_response_data <-
  tibble(
    # Patient categories, predefined and set terms
    "patient_category" = c(
      "SCA", 
      "CTAS 1",
      "CTAS 2",
      "CTAS 3",
      "CTAS 4",
      "CTAS 5"
    ),
    # Response time mandates, predefined and set numbers
    "plan_in_minutes" = c(
      6, 
      8,
      10,
      15,
      20,
      25
    ),
    # Performance mandates, predefined and set percentage
    "plan_in_percentage" = sample(
      x = 0.75,
      size = 6,
      replace = TRUE
    ),
    # Actual percentage of response times meeting performance plan, varying between 60% and 99%
    "performance_in_percentage" = sample(
      seq(0.6, 0.99, by = 0.001),
      size = 6,
      replace = TRUE
    ),
    # Sample year associated with data
    "response_time_standard_tear" = c(
      2021, 
      2021,
      2021,
      2021,
      2021,
      2021
    ),
)

simulated_response_data
```

```{r}
#### Simulate data based on Pre-Hospital Emergency Care Performance Metrics ####
# Link: https://open.toronto.ca/dataset/pre-hospital-emergency-care-performance-metrics
simulated_performance_data <-
  tibble(
    # List all years between 2014 and 2021, inclusive
    "year" = c(2014:2021),
    
    # Randomly select numbers between 200,000 and 300,000
    "ems_responses" = sample(
      x = c(200000:350000),
      size = 8,
      replace = TRUE
    )
)

simulated_performance_data
```

```{r}
#### Simulate data based on Paramedic Services Incident Data ####
# Link: https://open.toronto.ca/dataset/paramedic-services-incident-data
simulated_incident_data <-
  tibble(
    # Randomly pick dates between January 2, 2021 and December 31, 2021
    "dispatch_time" = sample(
      seq(as.Date('2021/01/01'), as.Date('2021/12/31'), by="day"), 
      size = 1000,
      replace = TRUE
    ),
    
    # Randomly pick an incident type 1000 times
    "incident_type" = sample(
      x = c(
        "Medical",
        "Motor Vehicle Accident",
        "Emergency Transfer",
        "Fire",
        "Airport Standby",
        "Other"
      ),
      size = 1000,
      replace = TRUE
    ),
    
    # Randomly select the number of units arriving on scene, between 1 and 22
    "units_arrived_at_scene" = sample(
      x = c(1:22),
      size = 1000,
      replace = TRUE
    )
  )

simulated_incident_data
```
